Move DrawableHitObject.
