+++ osu.Game.Modes.Catch/UI/CatchHitRenderer.cs
@@ -3,6 +3,7 @@
 
 using osu.Game.Modes.Catch.Objects;
 using osu.Game.Modes.Objects;
+using osu.Game.Modes.Objects.Drawables;
 using osu.Game.Modes.UI;
 
 namespace osu.Game.Modes.Catch.UI
